+++ R/obj_TEClassifierSequentialPrototype.R
@@ -91,6 +91,7 @@ TEClassifierSequentialPrototype <- R6::R6Class(
     #' @param metric_type `r get_param_doc_desc("metric_type")`
     #' @param embedding_dim `r get_param_doc_desc("embedding_dim")`
     #' @param projection_type `r get_param_doc_desc("projection_type")`
+    #' @param final_normalization_type `r get_param_doc_desc("final_normalization_type")`
     #' @return Function does nothing return. It modifies the current object.
     configure = function(name = NULL,
                          label = NULL,
@@ -146,7 +147,8 @@ TEClassifierSequentialPrototype <- R6::R6Class(
                          tf_parametrizations = "None",
                          tf_normalization_type = "LayerNorm",
                          tf_normalization_position = "Pre",
-                         tf_residual_type = "ResidualGate",
+                         tf_residual_type = "ResidualGate"                         ,
+                         final_normalization_type="PowerNorm",
                          embedding_dim = 2L) {
       arguments <- get_called_args(n = 1L)
       arguments$core_net_type <- "sequential"
@@ -213,7 +215,8 @@ TEClassifierSequentialPrototype <- R6::R6Class(
         tf_normalization_position = private$model_config$tf_normalization_position,
         tf_residual_type = private$model_config$tf_residual_type,
         embedding_dim = as.integer(private$model_config$embedding_dim),
-        core_net_type = private$model_config$core_net_type
+        core_net_type = private$model_config$core_net_type,
+        final_normalization_type=private$model_config$final_normalization_type
       )
       private$set_random_prototypes()
     },